requirements.txt
anyio==3.6.2
asgiref==3.5.2
bcrypt==3.2.0
cffi==1.15.1
click==8.1.3
dnspython==2.1.0
fastapi==0.70.0
fastapi-csrf-protect==0.2.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
motor==2.5.1
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.3.0
pymongo==3.13.0
python-decouple==3.3
six==1.16.0
sniffio==1.3.0
starlette==0.16.0
typing-extensions==4.4.0
uvicorn==0.14.0
zipp==3.11.0
